Atualização visual e inclusão da função EDITAR

+++ src/view/fund-components/ConsultFundItem.jsx
@@ -1,11 +1,15 @@
 import axios from "axios";
 import React from "react";
+import { Link } from "react-router-dom";
 
 function ConsultFundItem({ fundsList, fund, setFunds }) {
 
     function deleteFund(id) {
-        axios.delete(`http://localhost:3001/funds/${id}`)
-        setFunds(fundsList.filter(fund => fund.id !== id))
+        const resp = window.confirm(`Tem certeza que quer deletar "${fund.name}"?`)
+        if (resp === true) {
+          axios.delete(`http://localhost:3001/funds/${id}`)
+          setFunds(fundsList.filter(fund => fund.id !== id))
+        }
     }
     
   return (
@@ -14,8 +18,10 @@ function ConsultFundItem({ fundsList, fund, setFunds }) {
       <td className="table-content">{fund.cnpj}</td>
       <td className="table-content">{fund.creation_date ? fund.creation_date : "Data não preenchida"}</td>
       <td className="table-content">
-          <button>Editar</button>
-          <button onClick={() => deleteFund(fund.id)}>Deletar</button>
+          <Link to={`/funds/edit/${fund.id}`}>
+            <button className="btn-edit">Editar</button>
+          </Link>
+          <button className="btn-delete" onClick={() => deleteFund(fund.id)}>X</button>
       </td>
     </tr>
   );
